* FSM/dspchips/DSPChips.h: 	* Matilde/Tracker/BuzzInstrument.cpp: 	* Matilde/Tracker/BuzzSample.cpp: 	* Matilde/Tracker/SurfsDSPLib/SRF_Amp.h: 	* Matilde/Tracker/Tracker.cpp: 	  Build warning fixes.

+++ FSM/dspchips/DSPChips.h
@@ -1033,7 +1033,7 @@ inline int f2i(double d)
 	return *(int*) &tmp;
 #else
     // poor man's solution :)
-    return rint(d);
+    return (int)rint(d);
 #endif
 }
